requirements.txt
easydict==1.13
einops==0.8.0
h5py==3.11.0
hydra-core==1.3.2
imageio==2.34.1
matplotlib==3.9.0
numpy==1.26.4
positional-encodings==6.0.3
tqdm==4.66.4
vector-quantize-pytorch==1.14.24
wandb==0.17.0
metaworld==2.0.0
cmake==3.29.3
diffusers==0.28.0
pyinstrument==4.6.2
moviepy==1.0.3
tokenizers==0.19.1
torchtyping==0.1.4
ema-pytorch==0.5.2
natsort==8.4.0
opencv-python==4.6.0.66
thop==0.1.1-2209072238
bddl==1.0.1
future==0.18.2
gym==0.25.2
cloudpickle==2.1.0
robosuite<1.5
transformers
timm==0.9.16
